// libs/pmds-cdk/src/lib/pmds-cdk-text-input/models/pmds-cdk-text-input-tag-category-data.model.ts
////////PPMDS libraries
import { TPmdsDirectiveTooltipPosition } from "@pmhub/pmds-common";

export interface IPmdsCdkTextInputTagCategoryData {
    assetsFolder?: string;
    flagCode?: string;
    label?: string;
    showIcon?: boolean;
    tooltip?: string;
    tooltipPosition?: TPmdsDirectiveTooltipPosition;
}